feat: 헤더 drawer 위치 조정 및 구조 조정
