+++ src/shared/context/slide-drawer/DrawerStore.ts
@@ -1,0 +1,15 @@
+import { create } from 'zustand'
+
+interface DrawerState {
+  isOpen: boolean
+  open: () => void
+  close: () => void
+  toggle: () => void
+}
+
+export const useDrawerStore = create<DrawerState>(set => ({
+  isOpen: false,
+  open: () => set({ isOpen: true }),
+  close: () => set({ isOpen: false }),
+  toggle: () => set(state => ({ isOpen: !state.isOpen })),
+}))
