+++ src/shared/styles/LayoutWrapper.tsx
@@ -22,4 +22,6 @@ const Container = styled.div`
   position: relative;
   display: flex;
   flex-direction: column;
+
+  overflow-x: hidden;
 `
